refactor: move swap entity builder method to util file
